+++ crates/nostr-relay-builder/CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Breaking changes
 
 - Change `LocalRelay::new` constructor signature (https://github.com/rust-nostr/nostr/pull/1147)
+- Allow only a single `write` and `query` policy (https://github.com/rust-nostr/nostr/pull/1165)
 
 ### Changes